Split nested class for BlobStore
