+++ dbms/src/Storages/Page/V3/PageDirectoryFactory.h
@@ -15,6 +15,7 @@
 #pragma once
 
 #include <Storages/Page/PageDefines.h>
+#include <Storages/Page/V3/Blob/BlobStat.h>
 #include <Storages/Page/V3/BlobStore.h>
 #include <Storages/Page/V3/PageEntriesEdit.h>
 #include <Storages/Page/V3/WALStore.h>
@@ -54,7 +55,7 @@ public:
     PageDirectoryPtr createFromEdit(String storage_name, FileProviderPtr & file_provider, PSDiskDelegatorPtr & delegator, const PageEntriesEdit & edit);
 
     // just for test
-    PageDirectoryFactory & setBlobStats(BlobStore::BlobStats & blob_stats_)
+    PageDirectoryFactory & setBlobStats(BlobStats & blob_stats_)
     {
         blob_stats = &blob_stats_;
         return *this;
@@ -67,7 +68,7 @@ private:
         const PageDirectoryPtr & dir,
         const PageEntriesEdit::EditRecord & r);
 
-    BlobStore::BlobStats * blob_stats = nullptr;
+    BlobStats * blob_stats = nullptr;
 };
 
 } // namespace PS::V3
